Feat UD at detail page
- Split component to detail page
- Add UD feature to detail page map

by @kidongg #92
