+++ pages/detail/[id].tsx
@@ -6,20 +6,58 @@ import {
   postCounter,
 } from '@/api';
 import Seo from '@/components/Seo';
-import Image from 'next/image';
 import { useEffect, useState } from 'react';
-import { CustomOverlayMap, Map, MapMarker } from 'react-kakao-maps-sdk';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import styled from 'styled-components';
 import CommentList from '@/components/detail/CommentList';
 import FollowingButton from '@/components/detail/FollowingButton';
 import { authService } from '@/firebase';
 import DetailBox from '@/components/detail/DetailBox';
+import DetailMap from '@/components/detail/DetailMap';
 
 const Post = ({ id }: any) => {
-  const [isOpen, setIsOpen] = useState(false);
+  //* DetailMap state
+  //? category 클릭, 검색 시 map이동에 관한 통합 state
+  const [searchCategory, setSearchCategory]: any = useState('');
+  const [saveLatLng, setSaveLatLng]: any = useState([]);
+  const [saveAddress, setSaveAddress]: any = useState('');
+
+  // console.log('saveLatLng: ', saveLatLng);
+  // console.log('saveAddress: ', saveAddress);
+
+  //? 카테고리버튼 눌렀을 때 실행하는 함수
+  const [place, setPlace] = useState('');
+  const onClickEditTown = (e: any) => {
+    setPlace('');
+    setEditTown(e.target.innerText);
+    setSearchCategory(e.target.innerText);
+  };
+
+  //* DetailBox state
+  const [dropdownToggle, setDropdownToggle] = useState(false);
+  const [editTitle, setEditTitle] = useState('');
+  const [editContent, setEditContent] = useState('');
+  const [editCity, setEditCity] = useState('');
+  const [editTown, setEditTown] = useState('');
+  const [imageUpload, setImageUpload]: any = useState(null);
+
+  let editState = {
+    title: editTitle,
+    content: editContent,
+    city: editCity,
+    town: editTown,
+    imgUrl: '',
+    lat: saveLatLng.Ma,
+    long: saveLatLng.La,
+    address: saveAddress,
+  };
+
+  //* input 토글
+  const [inputToggle, setInputToggle] = useState(false);
+
   //* collection 저장 state
   const [isCollect, setIsCollect] = useState(true);
+  const [isOpen, setIsOpen] = useState(false);
 
   //* useQuery 사용해서 데이터 불러오기
   const {
@@ -115,78 +153,41 @@ const Post = ({ id }: any) => {
         })
         .map((item: any) => (
           <div key={item.id}>
-            <Map // 지도를 표시할 Container
-              center={{
-                // 지도의 중심좌표
-                lat: item.lat,
-                lng: item.long,
-              }}
-              style={{
-                // 지도의 크기
-                width: '100%',
-                height: '450px',
-              }}
-              level={6} // 지도의 확대 레벨
-            >
-              <MapMarker
-                position={{
-                  // 지도의 중심좌표
-                  lat: item.lat,
-                  lng: item.long,
-                }}
-                onClick={() => setIsOpen(true)}
-              />
-              {isOpen && (
-                <CustomOverlayMap
-                  position={{
-                    // 지도의 중심좌표
-                    lat: item.lat,
-                    lng: item.long,
-                  }}
-                >
-                  <StOverLayWrap>
-                    <StOverLayInfo>
-                      <StOverLayTitle>
-                        {item.title}
-                        <StOverLayClose
-                          className="close"
-                          onClick={() => setIsOpen(false)}
-                          title="닫기"
-                        ></StOverLayClose>
-                      </StOverLayTitle>
-                      <StOverLayBody>
-                        <StOverLayImg>
-                          <Image
-                            src={item.imgUrl}
-                            alt="image"
-                            height={100}
-                            width={100}
-                          />
-                        </StOverLayImg>
-                        <StOverLayDesc>
-                          <StOverLayEllipsis>{item.address}</StOverLayEllipsis>
-                          <StOverLayCounter>
-                            조회수 : {item.clickCounter}
-                          </StOverLayCounter>
-                          <div>
-                            <a
-                              href="https://www.kakaocorp.com/main"
-                              target="_blank"
-                              className="link"
-                              rel="noreferrer"
-                            >
-                              페이지이동(기능추가예정)
-                            </a>
-                          </div>
-                        </StOverLayDesc>
-                      </StOverLayBody>
-                    </StOverLayInfo>
-                  </StOverLayWrap>
-                </CustomOverlayMap>
-              )}
-            </Map>
             <StDetailBox>
-              <DetailBox item={item} />
+              <DetailMap
+                item={item}
+                setIsOpen={setIsOpen}
+                isOpen={isOpen}
+                inputToggle={inputToggle}
+                searchCategory={searchCategory}
+                saveLatLng={saveLatLng}
+                setSaveLatLng={setSaveLatLng}
+                saveAddress={saveAddress}
+                setSaveAddress={setSaveAddress}
+                setPlace={setPlace}
+                place={place}
+              />
+              <DetailBox
+                item={item}
+                inputToggle={inputToggle}
+                setInputToggle={setInputToggle}
+                setImageUpload={setImageUpload}
+                imageUpload={imageUpload}
+                editTitle={editTitle}
+                editContent={editContent}
+                editCity={editCity}
+                editTown={editTown}
+                editState={editState}
+                setEditTitle={setEditTitle}
+                setEditContent={setEditContent}
+                setEditCity={setEditCity}
+                setEditTown={setEditTown}
+                setDropdownToggle={setDropdownToggle}
+                dropdownToggle={dropdownToggle}
+                onClickEditTown={onClickEditTown}
+                setSaveLatLng={setSaveLatLng}
+                setSaveAddress={setSaveAddress}
+              />
               <FollowingButton item={item} />
             </StDetailBox>
           </div>
@@ -202,104 +203,6 @@ const StDetailBox = styled.div`
   border: 1px solid black;
   padding: 0px;
   margin: 10px;
-`;
-
-const StOverLayWrap = styled.div`
-  position: absolute;
-  left: 0;
-  bottom: 50px;
-  width: 288px;
-  height: 132px;
-  margin-left: -144px;
-  text-align: left;
-  overflow: hidden;
-  font-size: 12px;
-  font-family: 'Malgun Gothic', dotum, '돋움', sans-serif;
-  line-height: 1.5;
-  padding: 0;
-  margin: 20;
-`;
-
-const StOverLayInfo = styled.div`
-  width: 286px;
-  height: 120px;
-  border-radius: 5px;
-  border-bottom: 2px solid #ccc;
-  border-right: 1px solid #ccc;
-  overflow: hidden;
-  background: #fff;
-
-  :nth-child(1) {
-    border: 0;
-    box-shadow: 0px 1px 2px #888;
-  }
-  ::after {
-    content: '';
-    position: absolute;
-    margin-left: -12px;
-    left: 50%;
-    bottom: 0;
-    width: 22px;
-    height: 12px;
-    background: url('https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/vertex_white.png');
-  }
-  .link {
-    color: #5085bb;
-  }
-`;
-
-const StOverLayTitle = styled.div`
-  padding: 5px 0 0 10px;
-  height: 30px;
-  background: #eee;
-  border-bottom: 1px solid #ddd;
-  font-size: 18px;
-  font-weight: bold;
-`;
-
-const StOverLayClose = styled.div`
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  color: #888;
-  width: 17px;
-  height: 17px;
-  background: url('https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/overlay_close.png');
-  :hover {
-    cursor: pointer;
-  }
-`;
-
-const StOverLayBody = styled.div`
-  position: relative;
-  overflow: hidden;
-  cursor: inherit;
-`;
-
-const StOverLayDesc = styled.div`
-  position: relative;
-  margin: 13px 0 0 90px;
-  height: 75px;
-`;
-const StOverLayEllipsis = styled.div`
-  overflow: hidden;
-  /* text-overflow: ellipsis; */
-  white-space: nowrap;
-`;
-const StOverLayCounter = styled.div`
-  font-size: 11px;
-  color: #888;
-  margin-top: -2px;
-`;
-const StOverLayImg = styled.div`
-  position: absolute;
-  top: 6px;
-  left: 5px;
-  width: 73px;
-  height: 71px;
-  border: 1px solid #ddd;
-  color: #888;
-  overflow: hidden;
 `;
 
 //* SSR방식으로 server에서 id 값 보내기
